replace count.py with brie_count_droplet.py
