+++ brie/bin/__init__.py
@@ -1,4 +1,4 @@
-from .count import count
+from .count import smartseq_count, droplet_count
 from .quant import quant
 
 __all__ = ["count", "detect"]
